chore: move edit-settings to settings
